refactor: Move NetworkProxy profile generation logic into a new subdirectory
